Big Update (Update 3)

+++ src/main/java/cn/nukkit/inventory/ContainerInventory.java
@@ -1,13 +1,13 @@
 package cn.nukkit.inventory;
 
+import java.util.Map;
+
 import cn.nukkit.Player;
 import cn.nukkit.item.Item;
 import cn.nukkit.math.NukkitMath;
 import cn.nukkit.math.Vector3;
 import cn.nukkit.network.protocol.ContainerClosePacket;
 import cn.nukkit.network.protocol.ContainerOpenPacket;
-
-import java.util.Map;
 
 /**
  * author: MagicDroidX
@@ -34,8 +34,8 @@ public abstract class ContainerInventory extends BaseInventory {
     public void onOpen(Player who) {
         super.onOpen(who);
         ContainerOpenPacket pk = new ContainerOpenPacket();
-        pk.windowid = (byte) who.getWindowId(this);
-        pk.type = (byte) this.getType().getNetworkType();
+        pk.windowid = who.getWindowId(this);
+        pk.type = this.getType().getNetworkType();
         InventoryHolder holder = this.getHolder();
         if (holder instanceof Vector3) {
             pk.x = (int) ((Vector3) holder).getX();
@@ -53,7 +53,7 @@ public abstract class ContainerInventory extends BaseInventory {
     @Override
     public void onClose(Player who) {
         ContainerClosePacket pk = new ContainerClosePacket();
-        pk.windowid = (byte) who.getWindowId(this);
+        pk.windowid = who.getWindowId(this);
         who.dataPacket(pk);
         super.onClose(who);
     }
